wayland: switch to xdg interface
The xdg_shell interface replaces the old wl_shell interface. It provides more
features and is simpler to use.

The protocol file is distributed by mpv and the header and binding code is
generated from the protocol file. This is for some reason recommended.

If no xdg_shell is avaible the window won't show up.